refactor: use u64 instead of i128

+++ src/cli.rs
@@ -3,16 +3,16 @@ use structopt::StructOpt;
 #[derive(Debug, StructOpt)]
 pub struct Cli {
     /// minimum value as an integar
-    pub min: i128,
+    pub min: u64,
     /// minimum value as an integar
-    pub max: i128,
+    pub max: u64,
     #[structopt(short, long, parse(try_from_str = parse_key_val))]
     /// create custom matchers
     ///
     /// syntax: replace=number
     ///
     /// example: fizzbuzz 0 10 -d fizz=3 buzz=5 bazz=7
-    pub defines: Option<Vec<(String, i128)>>,
+    pub defines: Option<Vec<(String, u64)>>,
 }
 
 fn parse_key_val<T, U>(s: &str) -> Result<(T, U), Box<dyn Error>>
